Split asset management tests into two files
