+++ packages/fmg-nitro-adjudicator/test/nitro-erc20-management.test.ts
@@ -144,352 +144,6 @@ const guarantorCommitment = CountingApp.createCommitment.app({
   turnNum: 6,
 });
 
-describe('Nitro (ETH management)', () => {
-  let networkId;
-
-  // ETH management
-  // ========================
-
-  beforeAll(async () => {
-    networkId = (await provider.getNetwork()).chainId;
-    const libraryAddress = NitroArtifact.networks[networkId].address;
-    nitro = new ethers.Contract(libraryAddress, NitroArtifact.abi, signer0);
-  });
-
-  describe('Depositing ETH (msg.value = amount , expectedHeld = 0)', () => {
-    let tx;
-    let receipt;
-    const randomAddress = ethers.Wallet.createRandom().address;
-
-    it('Transaction succeeds', async () => {
-      tx = await nitro.deposit(randomAddress, 0, DEPOSIT_AMOUNT, AddressZero, {
-        value: DEPOSIT_AMOUNT,
-      });
-      receipt = await tx.wait();
-      await expect(receipt.status).toEqual(1);
-    });
-
-    it('Updates holdings', async () => {
-      const allocatedAmount = await nitro.holdings(randomAddress, AddressZero);
-      await expect(allocatedAmount).toEqual(DEPOSIT_AMOUNT);
-    });
-
-    it('Fires a deposited event', async () => {
-      await expectEvent(receipt, 'Deposited', {
-        destination: randomAddress,
-        amountDeposited: bigNumberify(DEPOSIT_AMOUNT),
-      });
-    });
-  });
-
-  describe('Depositing ETH (msg.value = amount, expectedHeld > holdings)', () => {
-    let tx;
-    const randomAddress = ethers.Wallet.createRandom().address;
-
-    it('Reverts', async () => {
-      tx = nitro.deposit(randomAddress, 10, DEPOSIT_AMOUNT, AddressZero, {
-        value: DEPOSIT_AMOUNT,
-      });
-      await expectRevert(() => tx, 'Deposit: holdings[destination][token] is less than expected');
-    });
-  });
-
-  describe('Depositing ETH (msg.value = amount, expectedHeld + amount < holdings)', () => {
-    let tx1;
-    let tx2;
-    let receipt;
-    let balanceBefore;
-    const randomAddress = ethers.Wallet.createRandom().address;
-
-    beforeAll(async () => {
-      tx1 = await nitro.deposit(randomAddress, 0, DEPOSIT_AMOUNT.mul(2), AddressZero, {
-        value: DEPOSIT_AMOUNT.mul(2),
-      });
-      await tx1.wait();
-      balanceBefore = await signer0.getBalance();
-      tx2 = await nitro.deposit(randomAddress, 0, DEPOSIT_AMOUNT, AddressZero, {
-        value: DEPOSIT_AMOUNT,
-      });
-      receipt = await tx2.wait();
-    });
-    it('Emits Deposit of 0 event ', async () => {
-      await expectEvent(receipt, 'Deposited', {
-        destination: randomAddress,
-        amountDeposited: bigNumberify(0),
-      });
-    });
-    it('Refunds entire deposit', async () => {
-      await expect(await signer0.getBalance()).toEqual(balanceBefore); // TODO handle gas fees
-    });
-  });
-
-  describe('Depositing ETH (msg.value = amount,  amount < holdings < amount + expectedHeld)', () => {
-    let tx1;
-    let tx2;
-    let receipt;
-    let balanceBefore;
-    const randomAddress = ethers.Wallet.createRandom().address;
-
-    beforeAll(async () => {
-      tx1 = await nitro.deposit(randomAddress, 0, DEPOSIT_AMOUNT.mul(11), AddressZero, {
-        value: DEPOSIT_AMOUNT.mul(11),
-      });
-      await tx1.wait();
-      balanceBefore = await signer0.getBalance();
-      tx2 = await nitro.deposit(
-        randomAddress,
-        DEPOSIT_AMOUNT.mul(10),
-        DEPOSIT_AMOUNT.mul(2),
-        AddressZero,
-        {
-          value: DEPOSIT_AMOUNT.mul(2),
-        },
-      );
-      receipt = await tx2.wait();
-    });
-    it('Emits Deposit event (partial) ', async () => {
-      await expectEvent(receipt, 'Deposited', {
-        destination: randomAddress,
-        amountDeposited: DEPOSIT_AMOUNT.mul(1),
-      });
-    });
-    it('Partial refund', async () => {
-      await expect(Number(await signer0.getBalance())).toBeGreaterThan(
-        Number(balanceBefore.sub(DEPOSIT_AMOUNT.mul(2))),
-      ); // TODO compute precisely, taking actual gas fees into account
-    });
-  });
-
-  describe('Withdrawing ETH (signer = participant, holdings[participant][0x] = 2 * amount)', () => {
-    let tx1;
-    let tx2;
-    let receipt;
-    let beforeBalance;
-    let allocatedAtStart;
-    const WITHDRAWAL_AMOUNT = DEPOSIT_AMOUNT;
-
-    beforeAll(async () => {
-      tx1 = await nitro.deposit(alice.address, 0, DEPOSIT_AMOUNT.mul(2), AddressZero, {
-        value: DEPOSIT_AMOUNT.mul(2),
-      });
-      await tx1.wait();
-      allocatedAtStart = await nitro.holdings(alice.address, AddressZero);
-      beforeBalance = await provider.getBalance(aliceDest.address);
-      tx2 = await withdraw(alice, aliceDest.address, alice, WITHDRAWAL_AMOUNT);
-      receipt = await tx2.wait();
-    });
-
-    it('Transaction succeeds', async () => {
-      await expect(receipt.status).toEqual(1);
-    });
-
-    it('Destination balance increases', async () => {
-      await expect(await provider.getBalance(aliceDest.address)).toEqual(
-        beforeBalance.add(WITHDRAWAL_AMOUNT),
-      );
-    });
-
-    it('holdings[participant][0x] decreases', async () => {
-      await expect(await nitro.holdings(alice.address, AddressZero)).toEqual(
-        allocatedAtStart.sub(WITHDRAWAL_AMOUNT),
-      );
-    });
-  });
-
-  describe('Withdrawing ETH (signer =/= partcipant, holdings[participant][0x] = amount)', () => {
-    let tx1;
-    let tx2;
-    let beforeBalance;
-    let allocatedAtStart;
-    const WITHDRAWAL_AMOUNT = DEPOSIT_AMOUNT;
-
-    beforeAll(async () => {
-      tx1 = await nitro.deposit(alice.address, 0, DEPOSIT_AMOUNT.mul(2), AddressZero, {
-        value: DEPOSIT_AMOUNT.mul(2),
-      });
-      await tx1.wait();
-      allocatedAtStart = await nitro.holdings(alice.address, AddressZero);
-      beforeBalance = await provider.getBalance(aliceDest.address);
-      tx2 = withdraw(alice, aliceDest.address, bob, WITHDRAWAL_AMOUNT);
-    });
-
-    it('Reverts', async () => {
-      await expectRevert(() => tx2, 'Withdraw: not authorized by participant');
-    });
-  });
-
-  describe('Withdrawing ETH (signer = partcipant, holdings[participant][0x] < amount)', () => {
-    let tx2;
-    const WITHDRAWAL_AMOUNT = DEPOSIT_AMOUNT;
-
-    beforeAll(async () => {
-      tx2 = withdraw(bob, aliceDest.address, bob, WITHDRAWAL_AMOUNT);
-    });
-
-    it('Reverts', async () => {
-      await expectRevert(() => tx2, 'Withdraw: overdrawn');
-    });
-  });
-
-  describe('Withdrawing ETH (signer = partcipant, holdings[participant][0x] > amount)', () => {
-    let tx2;
-    const WITHDRAWAL_AMOUNT = DEPOSIT_AMOUNT;
-
-    beforeAll(async () => {
-      tx2 = withdraw(bob, aliceDest.address, bob, WITHDRAWAL_AMOUNT);
-    });
-
-    it('Reverts', async () => {
-      await expectRevert(() => tx2, 'Withdraw: overdrawn');
-    });
-  });
-
-  describe('Transferring ETH (outcome = final, holdings[fromChannel] > outcomes[fromChannel].destination', () => {
-    let allocatedToChannel;
-    let allocatedToAlice;
-    beforeAll(async () => {
-      const amountHeldAgainstLedgerChannel = await nitro.holdings(
-        getChannelID(ledgerChannel),
-        AddressZero,
-      );
-      await nitro.deposit(
-        getChannelID(ledgerChannel),
-        amountHeldAgainstLedgerChannel,
-        DEPOSIT_AMOUNT,
-        AddressZero,
-        { value: DEPOSIT_AMOUNT },
-      );
-      const amountHeldAgainstGuarantorChannel = await nitro.holdings(
-        guarantor.address,
-        AddressZero,
-      );
-      await nitro.deposit(
-        guarantor.address,
-        amountHeldAgainstGuarantorChannel,
-        DEPOSIT_AMOUNT,
-        AddressZero,
-        { value: DEPOSIT_AMOUNT },
-      );
-
-      const allocationOutcome = {
-        destination: [alice.address, bob.address],
-        allocation,
-        finalizedAt: ethers.utils.bigNumberify(1),
-        challengeCommitment: getEthersObjectForCommitment(commitment0),
-        token: [AddressZero, AddressZero],
-      };
-      const tx = await nitro.setOutcome(getChannelID(ledgerChannel), allocationOutcome);
-      await tx.wait();
-
-      allocatedToChannel = await nitro.holdings(getChannelID(ledgerChannel), AddressZero);
-      allocatedToAlice = await nitro.holdings(alice.address, AddressZero);
-    });
-
-    it('Nitro.transfer tx succeeds', async () => {
-      const tx1 = await nitro.transfer(
-        getChannelID(ledgerChannel),
-        alice.address,
-        allocation[0],
-        AddressZero,
-      );
-      const receipt1 = await tx1.wait();
-      await expect(receipt1.status).toEqual(1);
-    });
-
-    it('holdings[to][0x] increases', async () => {
-      expect(await nitro.holdings(alice.address, AddressZero)).toEqual(
-        allocatedToAlice.add(allocation[0]),
-      );
-    });
-
-    it('holdings[from][0x] decreases', async () => {
-      expect(await nitro.holdings(getChannelID(ledgerChannel), AddressZero)).toEqual(
-        allocatedToChannel.sub(allocation[0]),
-      );
-    });
-  });
-
-  describe('Claiming ETH from a Guarantor', () => {
-    const finalizedAt = ethers.utils.bigNumberify(1);
-    let recipient;
-    const claimAmount = 2;
-    let expectedOutcome;
-    let startBal;
-    let startBalRecipient;
-
-    beforeAll(async () => {
-      recipient = bob.address;
-      const guarantee = {
-        destination: [bob.address, alice.address],
-        allocation: [],
-        finalizedAt,
-        challengeCommitment: getEthersObjectForCommitment(guarantorCommitment),
-        token: [AddressZero, AddressZero],
-      };
-      const allocationOutcome = {
-        destination: [alice.address, bob.address],
-        allocation,
-        finalizedAt,
-        challengeCommitment: getEthersObjectForCommitment(guarantorCommitment),
-        token: [AddressZero, AddressZero],
-      };
-      await (await nitro.setOutcome(guarantor.address, guarantee)).wait();
-      await (await nitro.setOutcome(getChannelID(ledgerChannel), allocationOutcome)).wait();
-
-      // TODO reinstate these ?
-      // expect(
-      //   getOutcomeFromParameters(await nitro.getOutcome(getChannelID(ledgerChannel))),
-      // ).toMatchObject(allocationOutcome);
-      // expect(getOutcomeFromParameters(await nitro.getOutcome(guarantor.address))).toMatchObject(
-      //   guarantee,
-      // );
-
-      startBal = 5;
-      await (await nitro.deposit(guarantor.address, 0, startBal, AddressZero, {
-        value: startBal,
-      })).wait();
-
-      // Other tests may have deposited into guarantor.address, but we
-      // ensure that the guarantor has at least 5 in holdings
-      startBal = await nitro.holdings(guarantor.address, AddressZero);
-      startBalRecipient = (await nitro.holdings(recipient, AddressZero)).toNumber();
-      const bAllocation = bigNumberify(bBal)
-        .sub(claimAmount)
-        .toHexString();
-      const allocationAfterClaim = [aBal, bAllocation];
-      expectedOutcome = {
-        destination: [alice.address, bob.address],
-        allocation: allocationAfterClaim,
-        finalizedAt: ethers.utils.bigNumberify(finalizedAt),
-        challengeCommitment: getEthersObjectForCommitment(guarantorCommitment),
-      };
-    });
-
-    it('Nitro.claim tx succeeds', async () => {
-      const tx1 = await nitro.claim(guarantor.address, recipient, claimAmount, AddressZero);
-      const receipt1 = await tx1.wait();
-      await expect(receipt1.status).toEqual(1);
-    });
-
-    it('New outcome registered', async () => {
-      const newOutcome = await nitro.getOutcome(getChannelID(ledgerChannel));
-      expect(getOutcomeFromParameters(newOutcome)).toMatchObject(expectedOutcome);
-    });
-
-    it('holdings[gurantor][0x] decreases', async () => {
-      expect(Number(await nitro.holdings(guarantor.address, AddressZero))).toEqual(
-        startBal - claimAmount,
-      );
-    });
-
-    it('holdings[recipient][0x] decreases', async () => {
-      expect(Number(await nitro.holdings(recipient, AddressZero))).toEqual(
-        startBalRecipient + claimAmount,
-      );
-    });
-  });
-});
-
 // ERC20 management
 // ========================
 let erc20;
@@ -498,6 +152,8 @@ let nitroAddress;
 describe('Nitro (ERC20 management)', () => {
   beforeAll(async () => {
     const networkId = (await provider.getNetwork()).chainId;
+    const libraryAddress = NitroArtifact.networks[networkId].address;
+    nitro = new ethers.Contract(libraryAddress, NitroArtifact.abi, signer0);
     erc20Address = ERC20Artifact.networks[networkId].address;
     erc20 = new ethers.Contract(erc20Address, ERC20Artifact.abi, signer0);
     nitroAddress = NitroArtifact.networks[networkId].address;
